Move tests alongside source code and consolidate internals
