+++ src/_mut_stack.ts
