+++ src/eval_test.ts
@@ -1,9 +1,25 @@
+/*
+ * Copyright 2022-2023 Josh Martinez
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import { expect } from "chai";
 import * as fc from "fast-check";
-import { cmb, Semigroup } from "../src/cmb.js";
-import { Eq, eq } from "../src/cmp.js";
-import { Eval } from "../src/eval.js";
-import { arbStr, expectLawfulSemigroup, tuple } from "./util.js";
+import { arbStr, expectLawfulSemigroup, tuple } from "./_test/utils.js";
+import { Semigroup, cmb } from "./cmb.js";
+import { Eq, eq } from "./cmp.js";
+import { Eval } from "./eval.js";
 
 describe("eval.js", () => {
     describe("Eval", () => {
