+++ src/maybe_test.ts
@@ -1,8 +1,21 @@
+/*
+ * Copyright 2022-2023 Josh Martinez
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import { expect } from "chai";
 import * as fc from "fast-check";
-import { cmb } from "../src/cmb.js";
-import { cmp, eq, Ordering } from "../src/cmp.js";
-import { Maybe } from "../src/maybe.js";
 import {
     arbNum,
     arbStr,
@@ -12,7 +25,10 @@ import {
     Num,
     Str,
     tuple,
-} from "./util.js";
+} from "./_test/utils.js";
+import { cmb } from "./cmb.js";
+import { cmp, eq, Ordering } from "./cmp.js";
+import { Maybe } from "./maybe.js";
 
 function nothing<T>(): Maybe<T> {
     return Maybe.nothing;
